rename to finite_checks
