+++ pytorch_lightning/trainer/training_loop.py
@@ -29,7 +29,7 @@ from pytorch_lightning.utilities import _TPU_AVAILABLE, AMPType, DeviceType, par
 from pytorch_lightning.utilities.distributed import rank_zero_info
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from pytorch_lightning.utilities.model_helpers import is_overridden
-from pytorch_lightning.utilities.nan import detect_nan_parameters
+from pytorch_lightning.utilities.finite_checks import detect_nan_parameters
 from pytorch_lightning.utilities.parsing import AttributeDict
 from pytorch_lightning.utilities.warnings import WarningCache
 
@@ -637,7 +637,7 @@ class TrainLoop:
 
             # check if loss or model weights are nan
             if self.trainer.terminate_on_nan:
-                self._check_nan(opt_closure_result.loss)
+                self._check_finite(opt_closure_result.loss)
 
             # track all the outputs across all steps
             batch_opt_idx = opt_idx if len(batch_outputs) > 1 else 0
@@ -679,7 +679,7 @@ class TrainLoop:
 
                     # check if loss or model weights are nan
                     if self.trainer.terminate_on_nan:
-                        self._check_nan(result.loss)
+                        self._check_finite(result.loss)
 
                 else:
                     self.warning_cache.warn("training_step returned None if it was on purpose, ignore this warning...")
@@ -690,7 +690,7 @@ class TrainLoop:
 
         return result
 
-    def _check_nan(self, loss: torch.Tensor) -> None:
+    def _check_finite(self, loss: torch.Tensor) -> None:
         if not torch.isfinite(loss).all():
             raise ValueError(f'The loss returned in `training_step` is {loss}.')
         model = self.trainer.lightning_module
